refactor: matching conventions
모든 내용을 컨밴션에 맞추어놓음.
